+++ eras/babbage/impl/src/Cardano/Ledger/Babbage.hs
@@ -28,14 +28,12 @@ import Cardano.Ledger.Alonzo (reapplyAlonzoTx)
 import Cardano.Ledger.Alonzo.Data (AlonzoAuxiliaryData (..), AuxiliaryData)
 import Cardano.Ledger.Alonzo.Scripts (AlonzoScript (..), Script)
 import Cardano.Ledger.Alonzo.TxInfo (ExtendedUTxO (..))
-import Cardano.Ledger.Alonzo.TxWitness (TxWitness (..))
 import Cardano.Ledger.Babbage.Era (BabbageEra)
 import Cardano.Ledger.Babbage.Genesis (AlonzoGenesis, extendPPWithGenesis)
 import Cardano.Ledger.Babbage.PParams (BabbagePParamsHKD (..))
-import Cardano.Ledger.Babbage.Rules (babbageMinUTxOValue)
+import Cardano.Ledger.Babbage.Rules ()
 import Cardano.Ledger.Babbage.Tx
-  ( AlonzoTx (..),
-    babbageInputDataHashes,
+  ( babbageInputDataHashes,
     babbageTxScripts,
     getDatumBabbage,
     minfee,
@@ -52,7 +50,6 @@ import Cardano.Ledger.Babbage.TxInfo (babbageTxInfo)
 import qualified Cardano.Ledger.Crypto as CC
 import Cardano.Ledger.Hashes (EraIndependentTxBody)
 import Cardano.Ledger.Keys (DSignable, Hash)
-import Cardano.Ledger.Serialization (mkSized)
 import qualified Cardano.Ledger.Shelley.API as API
 import Cardano.Ledger.Shelley.UTxO (UTxO (..))
 import Cardano.Ledger.ShelleyMA.Rules (consumed)
@@ -78,12 +75,6 @@ instance CC.Crypto c => API.CLI (BabbageEra c) where
   evaluateMinFee = minfee
 
   evaluateConsumed = consumed
-
-  addKeyWitnesses (AlonzoTx b ws aux iv) newWits = AlonzoTx b ws' aux iv
-    where
-      ws' = ws {txwitsVKey = Set.union newWits (txwitsVKey ws)}
-
-  evaluateMinLovelaceOutput pp out = babbageMinUTxOValue pp (mkSized out)
 
 instance CC.Crypto c => ExtendedUTxO (BabbageEra c) where
   txInfo = babbageTxInfo