+++ eras/conway/impl/src/Cardano/Ledger/Conway.hs
@@ -13,11 +13,9 @@ module Cardano.Ledger.Conway (ConwayEra) where
 import Cardano.Ledger.Alonzo (reapplyAlonzoTx)
 import Cardano.Ledger.Alonzo.Genesis (AlonzoGenesis)
 import Cardano.Ledger.Alonzo.TxInfo (ExtendedUTxO (..))
-import Cardano.Ledger.Alonzo.TxWitness (TxWitness (..))
-import Cardano.Ledger.Babbage.Rules (babbageMinUTxOValue)
+import Cardano.Ledger.Babbage.Rules ()
 import Cardano.Ledger.Babbage.Tx
-  ( AlonzoTx (..),
-    babbageInputDataHashes,
+  ( babbageInputDataHashes,
     babbageTxScripts,
     getDatumBabbage,
     minfee,
@@ -32,7 +30,6 @@ import Cardano.Ledger.Conway.TxOut (AlonzoEraTxOut (..), BabbageTxOut (..))
 import qualified Cardano.Ledger.Crypto as CC
 import Cardano.Ledger.Hashes (EraIndependentTxBody)
 import Cardano.Ledger.Keys (DSignable, Hash)
-import Cardano.Ledger.Serialization (mkSized)
 import qualified Cardano.Ledger.Shelley.API as API
 import Cardano.Ledger.Shelley.UTxO (UTxO (..))
 import Cardano.Ledger.ShelleyMA.Rules (consumed)
@@ -58,12 +55,6 @@ instance CC.Crypto c => API.CLI (ConwayEra c) where
   evaluateMinFee = minfee
 
   evaluateConsumed = consumed
-
-  addKeyWitnesses (AlonzoTx b ws aux iv) newWits = AlonzoTx b ws' aux iv
-    where
-      ws' = ws {txwitsVKey = Set.union newWits (txwitsVKey ws)}
-
-  evaluateMinLovelaceOutput pp out = babbageMinUTxOValue pp (mkSized out)
 
 instance CC.Crypto c => ExtendedUTxO (ConwayEra c) where
   txInfo = babbageTxInfo