+++ eras/shelley-ma/impl/src/Cardano/Ledger/Mary.hs
@@ -38,7 +38,7 @@ import Cardano.Ledger.Shelley.PParams (ShelleyPParamsUpdate)
 import qualified Cardano.Ledger.Shelley.PParams
 import Cardano.Ledger.ShelleyMA
 import Cardano.Ledger.ShelleyMA.AuxiliaryData (AuxiliaryData)
-import Cardano.Ledger.ShelleyMA.Rules.Utxo (consumed, scaledMinDeposit)
+import Cardano.Ledger.ShelleyMA.Rules.Utxo (consumed)
 import Cardano.Ledger.ShelleyMA.Rules.Utxow ()
 import Cardano.Ledger.ShelleyMA.Timelocks (Timelock)
 
@@ -58,13 +58,9 @@ instance CC.Crypto c => CLI (MaryEra c) where
 
   evaluateConsumed = consumed
 
-  addKeyWitnesses = addShelleyKeyWitnesses
-
-  evaluateMinLovelaceOutput pp (ShelleyTxOut _ v) = scaledMinDeposit v (_minUTxOValue pp)
-
 -- Self-Describing type synomyms
 
-type MaryEra c = ShelleyMAEra 'Mary c
+type MaryEra = ShelleyMAEra 'Mary
 
 type Self c = ShelleyMAEra 'Mary c