+++ eras/shelley-ma/impl/src/Cardano/Ledger/ShelleyMA/Rules/Utxo.hs
@@ -16,7 +16,6 @@ module Cardano.Ledger.ShelleyMA.Rules.Utxo
   ( ShelleyMAUTXO,
     ShelleyMAUtxoPredFailure (..),
     consumed,
-    scaledMinDeposit,
     validateOutsideValidityIntervalUTxO,
     validateValueNotConservedUTxO,
   )
@@ -75,52 +74,6 @@ import Lens.Micro
 import NoThunks.Class (NoThunks)
 import Numeric.Natural (Natural)
 import Validation
-
-{- The scaledMinDeposit calculation uses the minUTxOValue protocol parameter
-(passed to it as Coin mv) as a specification of "the cost of
-making a Shelley-sized UTxO entry", calculated here by "utxoEntrySizeWithoutVal + uint",
-using the constants in the "where" clause.
-In the case when a UTxO entry contains coins only (and the Shelley
-UTxO entry format is used - we will extend this to be correct for other
-UTxO formats shortly), the deposit should be exactly the minUTxOValue.
-This is the "inject (coin v) == v" case.
-Otherwise, we calculate the per-byte deposit by multiplying the minimum deposit (which is
-for the number of Shelley UTxO-entry bytes) by the size of a Shelley UTxO entry.
-This is the "(mv * (utxoEntrySizeWithoutVal + uint))" calculation.
-We then calculate the total deposit required for making a UTxO entry with a Val-class
-member v by dividing "(mv * (utxoEntrySizeWithoutVal + uint))" by the
-estimated total size of the UTxO entry containing v, ie by
-"(utxoEntrySizeWithoutVal + size v)".
-See the formal specification for details.
--}
-
--- This scaling function is right for UTxO, not EUTxO
---
-scaledMinDeposit :: (Val.Val v) => v -> Coin -> Coin
-scaledMinDeposit v (Coin mv)
-  | Val.inject (Val.coin v) == v = Coin mv -- without non-Coin assets, scaled deposit should be exactly minUTxOValue
-  -- The calculation should represent this equation
-  -- minValueParameter / coinUTxOSize = actualMinValue / valueUTxOSize
-  -- actualMinValue = (minValueParameter / coinUTxOSize) * valueUTxOSize
-  | otherwise = Coin $ max mv (coinsPerUTxOWord * (utxoEntrySizeWithoutVal + Val.size v))
-  where
-    -- lengths obtained from tracing on HeapWords of inputs and outputs
-    -- obtained experimentally, and number used here
-    -- units are Word64s
-    txoutLenNoVal = 14
-    txinLen = 7
-
-    -- unpacked CompactCoin Word64 size in Word64s
-    coinSize :: Integer
-    coinSize = 0
-
-    utxoEntrySizeWithoutVal :: Integer
-    utxoEntrySizeWithoutVal = 6 + txoutLenNoVal + txinLen
-
-    -- how much ada does a Word64 of UTxO space cost, calculated from minAdaValue PP
-    -- round down
-    coinsPerUTxOWord :: Integer
-    coinsPerUTxOWord = quot mv (utxoEntrySizeWithoutVal + coinSize)
 
 -- ==========================================================
 
@@ -226,7 +179,6 @@ utxoTransition ::
     HasField "_minfeeB" (PParams era) Natural,
     HasField "_keyDeposit" (PParams era) Coin,
     HasField "_poolDeposit" (PParams era) Coin,
-    HasField "_minUTxOValue" (PParams era) Coin,
     HasField "_maxTxSize" (PParams era) Natural
   ) =>
   TransitionRule (ShelleyMAUTXO era)
@@ -319,21 +271,18 @@ validateOutputTooBigUTxO (UTxO outputs) =
 --
 -- > ∀ txout ∈ txouts txb, getValue txout ≥ inject (scaledMinDeposit v (minUTxOValue pp))
 validateOutputTooSmallUTxO ::
-  ( HasField "_minUTxOValue" (PParams era) Coin,
-    EraTxOut era
-  ) =>
+  EraTxOut era =>
   PParams era ->
   UTxO era ->
   Test (ShelleyMAUtxoPredFailure era)
 validateOutputTooSmallUTxO pp (UTxO outputs) =
   failureUnless (null outputsTooSmall) $ OutputTooSmallUTxO outputsTooSmall
   where
-    minUTxOValue = getField @"_minUTxOValue" pp
     outputsTooSmall =
       filter
-        ( \out ->
-            let v = out ^. valueTxOutL
-             in Val.pointwise (<) v (Val.inject $ scaledMinDeposit v minUTxOValue)
+        ( \txOut ->
+            let v = txOut ^. valueTxOutL
+             in Val.pointwise (<) v (Val.inject $ getMinCoinTxOut pp txOut)
         )
         (Map.elems outputs)