+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/API/Wallet.hs
@@ -32,7 +32,10 @@ module Cardano.Ledger.Shelley.API.Wallet
 
     -- * Transaction helpers
     CLI (..),
+    addKeyWitnesses,
+    evaluateTransactionFee,
     evaluateTransactionBalance,
+    evaluateMinLovelaceOutput,
     addShelleyKeyWitnesses,
     -- -- * Ada pots
     AdaPots (..),
@@ -104,7 +107,7 @@ import Cardano.Ledger.Shelley.PoolRank
 import Cardano.Ledger.Shelley.RewardProvenance (RewardProvenance)
 import Cardano.Ledger.Shelley.Rewards (StakeShare (..))
 import Cardano.Ledger.Shelley.Rules.NewEpoch (calculatePoolDistr)
-import Cardano.Ledger.Shelley.Tx (ShelleyTx (..), ShelleyWitnesses, WitnessSetHKD (..))
+import Cardano.Ledger.Shelley.Tx (ShelleyTx (..))
 import Cardano.Ledger.Shelley.TxBody (PoolParams (..), ShelleyEraTxBody, WitVKey (..))
 import Cardano.Ledger.Shelley.UTxO (UTxO (..), coinConsumed)
 import Cardano.Ledger.Slot (epochInfoSize)
@@ -470,41 +473,42 @@ class
   -- Used for the default implentation of 'evaluateTransactionBalance'.
   evaluateConsumed :: PParams era -> UTxO era -> TxBody era -> Value era
 
-  addKeyWitnesses :: Tx era -> Set (WitVKey 'Witness (EraCrypto era)) -> Tx era
+-- | Evaluate the minimum lovelace that a given transaction output must contain.
+evaluateMinLovelaceOutput :: EraTxOut era => PParams era -> TxOut era -> Coin
+evaluateMinLovelaceOutput = getMinCoinTxOut
+{-# DEPRECATED evaluateMinLovelaceOutput "In favor of `getMinCoinTxOut`" #-}
 
-  -- | Evaluate the fee for a given transaction.
-  evaluateTransactionFee ::
-    -- | The current protocol parameters.
-    PParams era ->
-    -- | The transaction.
-    Tx era ->
-    -- | The number of key witnesses still to be added to the transaction.
-    Word ->
-    -- | The required fee.
-    Coin
-  evaluateTransactionFee pp tx numKeyWits =
-    evaluateMinFee @era pp tx'
-    where
-      sigSize = fromIntegral $ sizeSigDSIGN (Proxy @(DSIGN (EraCrypto era)))
-      dummySig =
-        fromRight
-          (error "corrupt dummy signature")
-          (decodeFullDecoder "dummy signature" decodeSignedDSIGN (serialize $ LBS.replicate sigSize 0))
-      vkeySize = fromIntegral $ sizeVerKeyDSIGN (Proxy @(DSIGN (EraCrypto era)))
-      dummyVKey w =
-        let padding = LBS.replicate paddingSize 0
-            paddingSize = vkeySize - LBS.length sw
-            sw = serialize w
-            keyBytes = serialize $ padding <> sw
-         in fromRight (error "corrupt dummy vkey") (decodeFull keyBytes)
-      dummyKeyWits = Set.fromList $
-        flip map [1 .. numKeyWits] $
-          \x -> WitVKey (dummyVKey x) dummySig
+addKeyWitnesses :: EraTx era => Tx era -> Set (WitVKey 'Witness (EraCrypto era)) -> Tx era
+addKeyWitnesses tx newWits = tx & witsTxL . addrWitsL %~ Set.union newWits
 
-      tx' = addKeyWitnesses @era tx dummyKeyWits
-
-  -- | Evaluate the minimum lovelace that a given transaction output must contain.
-  evaluateMinLovelaceOutput :: PParams era -> TxOut era -> Coin
+-- | Evaluate the fee for a given transaction.
+evaluateTransactionFee ::
+  forall era.
+  CLI era =>
+  -- | The current protocol parameters.
+  PParams era ->
+  -- | The transaction.
+  Tx era ->
+  -- | The number of key witnesses still to be added to the transaction.
+  Word ->
+  -- | The required fee.
+  Coin
+evaluateTransactionFee pp tx numKeyWits = evaluateMinFee pp tx'
+  where
+    sigSize = fromIntegral $ sizeSigDSIGN (Proxy @(DSIGN (EraCrypto era)))
+    dummySig =
+      fromRight
+        (error "corrupt dummy signature")
+        (decodeFullDecoder "dummy signature" decodeSignedDSIGN (serialize $ LBS.replicate sigSize 0))
+    vkeySize = fromIntegral $ sizeVerKeyDSIGN (Proxy @(DSIGN (EraCrypto era)))
+    dummyVKey w =
+      let padding = LBS.replicate paddingSize 0
+          paddingSize = vkeySize - LBS.length sw
+          sw = serialize w
+          keyBytes = serialize $ padding <> sw
+       in fromRight (error "corrupt dummy vkey") (decodeFull keyBytes)
+    dummyKeyWits = Set.fromList [WitVKey (dummyVKey x) dummySig | x <- [1 .. numKeyWits]]
+    tx' = addKeyWitnesses tx dummyKeyWits
 
 -- | Evaluate the difference between the value currently being consumed by
 -- a transaction and the number of lovelace being produced.
@@ -534,22 +538,17 @@ evaluateTransactionBalance pp u isNewPool txb =
 --------------------------------------------------------------------------------
 
 addShelleyKeyWitnesses ::
-  (EraTx era, Witnesses era ~ ShelleyWitnesses era) =>
+  (EraTx era, Tx era ~ ShelleyTx era) =>
   ShelleyTx era ->
   Set (WitVKey 'Witness (EraCrypto era)) ->
   ShelleyTx era
-addShelleyKeyWitnesses (ShelleyTx b ws aux) newWits = ShelleyTx b ws' aux
-  where
-    ws' = ws {addrWits = Set.union newWits (addrWits ws)}
+addShelleyKeyWitnesses = addKeyWitnesses
+{-# DEPRECATED addShelleyKeyWitnesses "In favor of 'addKeyWitnesses'" #-}
 
 instance CC.Crypto c => CLI (ShelleyEra c) where
   evaluateMinFee = minfee
 
   evaluateConsumed = coinConsumed
-
-  addKeyWitnesses = addShelleyKeyWitnesses
-
-  evaluateMinLovelaceOutput pp _out = _minUTxOValue pp
 
 --------------------------------------------------------------------------------
 -- CBOR instances