+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/Out.hs
@@ -15,32 +15,41 @@ module Cardano.Ledger.Api.Tx.Out
     -- * Babbage
     BabbageTxOut,
     BabbageEraTxOut (..),
-    setBabbageMinTxOut,
+    setMinCoinTxOut,
+    setMinCoinSizedTxOut,
   )
 where
 
 import Cardano.Ledger.Alonzo.TxBody (AlonzoEraTxOut (..), AlonzoTxOut)
-import Cardano.Ledger.Babbage.Rules (babbageMinUTxOValue)
 import Cardano.Ledger.Babbage.TxBody (BabbageEraTxOut (..), BabbageTxOut)
-import Cardano.Ledger.Coin
 import Cardano.Ledger.Core (EraTxOut (..), PParams, coinTxOutL)
 import Cardano.Ledger.Serialization
 import Cardano.Ledger.Shelley.TxBody (ShelleyTxOut)
-import GHC.Records
 import Lens.Micro
 
+-- | Same as `setMinCoinSizedTxOut`, except it doesn't require the size of the
+-- TxOut and will recompute it if needed. Initial amount is not important.
+setMinCoinTxOut :: EraTxOut era => PParams era -> TxOut era -> TxOut era
+setMinCoinTxOut pp = go
+  where
+    go txOut =
+      let curMinCoin = getMinCoinTxOut pp txOut
+          curCoin = txOut ^. coinTxOutL
+       in if curCoin == curMinCoin
+            then txOut
+            else go (txOut & coinTxOutL .~ curMinCoin)
+
 -- | This function will adjust the output's `Coin` value to the smallest amount
--- allowed by the UTXO rule starting with BabbageEra. Initial amount is not
--- important.
-setBabbageMinTxOut ::
-  (EraTxOut era, HasField "_coinsPerUTxOByte" (PParams era) Coin) =>
+-- allowed by the UTXO rule. Initial amount is not important.
+setMinCoinSizedTxOut ::
+  EraTxOut era =>
   PParams era ->
   Sized (TxOut era) ->
   Sized (TxOut era)
-setBabbageMinTxOut pp = go
+setMinCoinSizedTxOut pp = go
   where
     go txOut =
-      let curMinCoin = babbageMinUTxOValue pp txOut
+      let curMinCoin = getMinCoinSizedTxOut pp txOut
           curCoin = txOut ^. toSizedL coinTxOutL
        in if curCoin == curMinCoin
             then txOut